fix: Target config id not being propagated after signing in
