+++ src/hooks/redirection.ts
@@ -1,39 +1,11 @@
 import { useCallback } from 'react'
 import { useLocation } from 'react-router-dom'
-import { locations } from '../shared/locations'
+import { extractRedirectToFromSearchParameters, locations } from '../shared/locations'
 
 export const useAfterLoginRedirection = () => {
   const location = useLocation()
   const search = new URLSearchParams(location.search)
-
-  // Extract 'redirectTo' from current search parameters
-  let redirectToSearchParam = search.get('redirectTo')
-  try {
-    const state = search.get('state')
-    // Decode the state parameter to get the original 'redirectTo'
-    if (state) {
-      const stateRedirectToParam = atob(state)
-      const parsedRedirectTo = JSON.parse(stateRedirectToParam).customData
-      if (parsedRedirectTo) {
-        redirectToSearchParam = parsedRedirectTo ?? null
-      }
-    }
-  } catch (_) {
-    console.error("Can't decode state parameter")
-  }
-
-  // Initialize redirectTo with a default value
-  let redirectTo = locations.home()
-
-  // Decode 'redirectTo' if it exists
-  if (redirectToSearchParam) {
-    try {
-      redirectTo = decodeURIComponent(redirectToSearchParam)
-    } catch (error) {
-      console.error("Can't decode redirectTo parameter")
-    }
-  }
-
+  const redirectTo = extractRedirectToFromSearchParameters(search)
   let sanitizedRedirectTo = locations.home()
 
   try {
@@ -49,6 +21,12 @@ export const useAfterLoginRedirection = () => {
     // Check if the hostname matches to prevent open redirects
     if (redirectToURL.hostname !== window.location.hostname) {
       redirectToURL = new URL('/auth/invalidRedirection', window.location.origin)
+    }
+
+    // Add the targetConfigId to the redirect URL if it exists
+    const targetConfigId = search.get('targetConfigId')
+    if (targetConfigId && !redirectToURL.searchParams.has('targetConfigId')) {
+      redirectToURL.searchParams.append('targetConfigId', targetConfigId)
     }
 
     // Set the sanitized redirect URL
